# crates/storage-api/openapi-storage-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Storage API",
    "description": "",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/pubsub": {
      "post": {
        "tags": [
          "crate::routes::pubsub"
        ],
        "operationId": "pubsub_endpoint",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SNS-compatible response"
          }
        }
      }
    },
    "/queue": {
      "post": {
        "tags": [
          "crate::routes::queue"
        ],
        "operationId": "queue_endpoint",
        "requestBody": {
          "content": {
            "application/x-amz-json-1.0": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SQS-compatible response"
          }
        }
      }
    },
...
  },
...
}